pyproject.toml
[project]
name = "inventory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "markdown>=3.7",
    "pillow>=11.0.0",
    "ppf-datamatrix>=0.2",
    "python-dotenv>=1.0.1",
    "uuid>=1.30",
]

[tool.uv]
dev-dependencies = ["ruff>=0.6.9"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
preview = true
select = ["ALL"]
ignore = [
    "D",      # pydocstyle
    "CPY",    # flake8-copyright
    "T20",    # flake8-print
    "PGH003", # Use specific error codes when using noqa
]
